requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# AWS mocking
moto[all]>=4.2.0
boto3-stubs[essential]>=1.28.0

# Code quality
mypy>=1.5.0
black>=23.7.0
flake8>=6.1.0

# CDK testing
aws-cdk-lib>=2.100.0
constructs>=10.0.0